refactor: Present stream replacement in the settings as an option similar to `iOS` and `Android VR`
