+++ app/src/main/java/app/revanced/integrations/youtube/patches/spoof/requests/StoryboardRendererRequester.java
@@ -3,8 +3,8 @@ package app.revanced.integrations.youtube.patches.spoof.requests;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import app.revanced.integrations.shared.settings.BaseSettings;
+import app.revanced.integrations.youtube.patches.spoof.ClientType;
 import app.revanced.integrations.youtube.patches.spoof.StoryboardRenderer;
-import app.revanced.integrations.youtube.patches.spoof.SpoofClientPatch.ClientType;
 import app.revanced.integrations.youtube.requests.Requester;
 import app.revanced.integrations.shared.Logger;
 import app.revanced.integrations.shared.Utils;
